Struct map_gui::options::Options

pub struct Options {

    pub dev: bool,
    pub debug_all_agents: bool,
    pub traffic_signal_style: TrafficSignalStyle,
    pub color_scheme: ColorSchemeChoice,
    pub toggle_day_night_colors: bool,
    pub camera_angle: CameraAngle,
    pub show_building_driveways: bool,
    pub show_stop_signs: bool,
    pub show_crosswalks: bool,
    pub show_traffic_signal_icon: bool,
    pub simplify_basemap: bool,
    pub minimal_controls: bool,
    pub canvas_settings: CanvasSettings,
    pub time_increment: Duration,
    pub dont_draw_time_warp: bool,
    pub jump_to_delay: Duration,
    pub language: Option<String>,
    pub units: UnitFmt,

}

Options controlling the UI. Some of the options are common to all map-based apps, and some are specific to A/B Street.

Fields
dev: bool

Dev mode exposes experimental tools useful for debugging, but that’d likely confuse most players.

debug_all_agents: bool

Every time we draw, render all agents zoomed in. Extremely slow. Just used to flush out drawing bugs.

traffic_signal_style: TrafficSignalStyle

How traffic signals should be rendered.

color_scheme: ColorSchemeChoice

The color scheme for map elements, agents, and the UI.

toggle_day_night_colors: bool

Automatically change color_scheme based on simulation time to reflect day/night

camera_angle: CameraAngle

Draw buildings in different perspectives

show_building_driveways: bool

Draw building driveways.

show_stop_signs: bool

Draw stop signs.

show_crosswalks: bool

Draw crosswalks and unmarked crossings.

show_traffic_signal_icon: bool

If true, draw an icon for traffic signals both when zoomed and unzoomed. If false, color the intersection when unzoomed and render the signal’s current state when zoomed.

simplify_basemap: bool

If true, modify several basemap features to de-emphasize them: border intersections

minimal_controls: bool

When making a screen recording, enable this option to hide some UI elements

canvas_settings: CanvasSettings

widgetry options

time_increment: Duration

How much to advance the sim with one of the speed controls

dont_draw_time_warp: bool

When time warping, don’t draw to speed up simulation

jump_to_delay: Duration

The delay threshold to halt on when jumping to the next delay

language: Option<String>

Display roads and buildings in an alternate language, if possible. None means to use the OSM native name.

units: UnitFmt

How to render geometric units

Implementations

impl Options

pub fn load_or_default() -> Options

Restore previous options. If the file is missing or the format has changed, fall back to built-in defaults.