Struct map_gui::render::agents::AgentCache

pub struct AgentCache {
    pub unzoomed_agents: UnzoomedAgents,
    time: Option<Time>,
    agents_per_on: HashMap<Traversable, Vec<Box<dyn Renderable>>>,
    unzoomed: Option<(Time, UnzoomedAgents, QuadTree<AgentID>, Drawable)>,
}

Fields
unzoomed_agents: UnzoomedAgents

This is controlled almost entirely by the minimap panel. It has no meaning in edit mode.

time: Option<Time>agents_per_on: HashMap<Traversable, Vec<Box<dyn Renderable>>>unzoomed: Option<(Time, UnzoomedAgents, QuadTree<AgentID>, Drawable)>
Implementations

impl AgentCache

pub fn new() -> AgentCache

pub fn get(&self, on: Traversable) -> Vec<&dyn Renderable>

pub fn populate_if_needed(
&mut self,
on: Traversable,
map: &Map,
sim: &Sim,
cs: &ColorScheme,
prerender: &Prerender
)

pub fn calculate_unzoomed_agents<P: AsRef<Prerender>>(
&mut self,
prerender: &mut P,
map: &Map,
sim: &Sim,
cs: &ColorScheme
) -> &QuadTree<AgentID>

If the sim time has changed or the unzoomed agent filters have been modified, recalculate the quadtree and drawable for all unzoomed agents.

pub fn draw_unzoomed_agents(
&mut self,
g: &mut GfxCtx<'_>,
map: &Map,
sim: &Sim,
cs: &ColorScheme,
opts: &Options
)