Struct map_gui::render::map::DrawMap

pub struct DrawMap {

    pub roads: Vec<DrawRoad>,
    pub intersections: Vec<DrawIntersection>,
    pub buildings: Vec<DrawBuilding>,
    pub parking_lots: Vec<DrawParkingLot>,
    pub bus_stops: HashMap<TransitStopID, DrawTransitStop>,
    pub areas: Vec<DrawArea>,
    pub boundary_polygon: Drawable,
    pub draw_all_unzoomed_roads_and_intersections: Drawable,
    pub draw_all_buildings: Drawable,
    pub draw_all_building_outlines: Drawable,
    pub draw_all_unzoomed_parking_lots: Drawable,
    pub draw_all_areas: Drawable,
    pub zorder_range: (isize, isize),
    pub show_zorder: isize,
    quadtree: QuadTree<ID>,

}

Fields
roads: Vec<DrawRoad>intersections: Vec<DrawIntersection>buildings: Vec<DrawBuilding>parking_lots: Vec<DrawParkingLot>bus_stops: HashMap<TransitStopID, DrawTransitStop>areas: Vec<DrawArea>boundary_polygon: Drawabledraw_all_unzoomed_roads_and_intersections: Drawabledraw_all_buildings: Drawabledraw_all_building_outlines: Drawabledraw_all_unzoomed_parking_lots: Drawabledraw_all_areas: Drawablezorder_range: (isize, isize)show_zorder: isizequadtree: QuadTree<ID>
Implementations

impl DrawMap

pub fn new( ctx: &mut EventCtx<'_>, map: &Map, opts: &Options, cs: &ColorScheme, timer: &mut Timer<'_> ) -> DrawMap

pub fn regenerate_buildings( ctx: &EventCtx<'_>, map: &Map, cs: &ColorScheme, opts: &Options, timer: &mut Timer<'_> ) -> (Vec<DrawBuilding>, Drawable, Drawable)

pub fn regenerate_parking_lots( ctx: &EventCtx<'_>, map: &Map, cs: &ColorScheme, opts: &Options ) -> (Vec<DrawParkingLot>, Drawable)

pub fn regenerate_unzoomed_layer( ctx: &EventCtx<'_>, map: &Map, cs: &ColorScheme, opts: &Options, timer: &mut Timer<'_> ) -> Drawable

pub fn get_r(&self, id: RoadID) -> &DrawRoad

pub fn get_l(&self, id: LaneID) -> &DrawLane

pub fn get_i(&self, id: IntersectionID) -> &DrawIntersection

pub fn get_b(&self, id: BuildingID) -> &DrawBuilding

pub fn get_pl(&self, id: ParkingLotID) -> &DrawParkingLot

pub fn get_ts(&self, id: TransitStopID) -> &DrawTransitStop

pub fn get_a(&self, id: AreaID) -> &DrawArea

pub fn get_obj<'a>(&self, id: ID) -> &dyn Renderable

pub fn get_matching_objects(&self, bounds: Bounds) -> Vec<ID>

Unsorted, unexpanded, raw result.

pub fn get_renderables_back_to_front( &self, bounds: Bounds, map: &Map ) -> Vec<&dyn Renderable>

A simple variation of the one in game that shows all layers, ignores dynamic agents.

pub fn zoomed_batch(ctx: &EventCtx<'_>, app: &dyn AppLike) -> GeomBatch

Build a single gigantic GeomBatch to render the entire map when zoomed in. Likely messes up Z-ordering.

pub fn recreate_intersection(&mut self, i: IntersectionID, map: &Map)

pub fn recreate_road(&mut self, road: &Road, map: &Map)

pub fn free_memory(&mut self)