Struct map_gui::simple_app::SimpleApp

pub struct SimpleApp<T> {
    pub map: Map,
    pub draw_map: DrawMap,
    pub cs: ColorScheme,
    pub opts: Options,
    pub current_selection: Option<ID>,
    pub session: T,
    pub time: Time,
}

Simple app state that just renders a static map, without any dynamic agents on the map.

Fields
map: Mapdraw_map: DrawMapcs: ColorSchemeopts: Optionscurrent_selection: Option<ID>session: T

Custom per-app state can be stored here

time: Time

If desired, this can be advanced to render traffic signals changing.

Implementations

impl<T: 'static> SimpleApp<T>

pub fn new<F: 'static + Fn(&mut EventCtx<'_>, &mut SimpleApp<T>) -> Vec<Box<dyn State<SimpleApp<T>>>>>(
ctx: &mut EventCtx<'_>,
opts: Options,
session: T,
init_states: F
) -> (SimpleApp<T>, Vec<Box<dyn State<SimpleApp<T>>>>)

pub fn draw_unzoomed(&self, g: &mut GfxCtx<'_>)

pub fn draw_zoomed(&self, g: &mut GfxCtx<'_>, opts: DrawOptions)

pub fn recalculate_current_selection(&mut self, ctx: &EventCtx<'_>)

Assumes some defaults.

pub fn mouseover_unzoomed_roads_and_intersections(
&self,
ctx: &EventCtx<'_>
) -> Option<ID>

pub fn mouseover_unzoomed_buildings(&self, ctx: &EventCtx<'_>) -> Option<ID>

Only select buildings, and work whether zoomed in or not.

fn calculate_current_selection(
&self,
ctx: &EventCtx<'_>,
unzoomed_roads_and_intersections: bool,
unzoomed_buildings: bool
) -> Option<ID>