Struct map_gui::SimpleAppArgs

pub struct SimpleAppArgs {
    pub map_path: Option<String>,
    pub cam: Option<String>,
    pub dev: bool,
    pub color_scheme: Option<ColorSchemeChoice>,
    pub minimal_controls: bool,
    pub scale_factor: Option<f64>,
}

Fields
map_path: Option<String>

Path to a map to initially load. If not provided, load the last map used or a fixed default.

cam: Option<String>

Initially position the camera here. The format is an OSM-style zoom/lat/lon string (https://wiki.openstreetmap.org/wiki/Browsing#Other_URL_tricks).

dev: bool

Dev mode exposes experimental tools useful for debugging, but that’d likely confuse most players.

color_scheme: Option<ColorSchemeChoice>

The color scheme for map elements, agents, and the UI.

minimal_controls: bool

When making a screen recording, enable this option to hide some UI elements

scale_factor: Option<f64>

Override the monitor’s auto-detected scale factor

Implementations

impl SimpleAppArgs

pub fn override_options(&self, opts: &mut Options)

Options are passed in by each app, usually seeded with defaults or from a config file. For the few options that we allow to be specified by command-line, overwrite the values.

pub fn update_widgetry_settings(&self, settings: Settings) -> Settings

pub fn map_name(&self) -> MapName