Module map_gui::tools

Assorted tools and UI states that're useful for applications built to display maps.

Modules

city_picker

colors

heatmap

minimap

navigate

turn_explorer

ui
Generic UI tools. Some of this should perhaps be lifted to widgetry.

updater

Structs

ChooseSomething
Choose something from a menu, then feed the answer to a callback.

CityPicker
Lets the player switch maps.

ColorDiscrete

ColorLegend

ColorNetwork

ColorScale

DivergingScale

Grid

HeatmapOptions

Minimap

Navigator

PopupMsg
Display a message dialog.

PromptInput
Prompt for arbitrary text input, then feed the answer to a callback.

TurnExplorer
A tool to explore all of the turns from a single lane.

Traits

MinimapControls
Customize the appearance and behavior of a minimap.

Functions

grey_out_map
Make it clear the map can't be interacted with right now.

loading_tips

make_heatmap

nice_map_name

open_browser