Enum map_gui::tools::Executable

pub enum Executable {
    ABStreet,
    FifteenMin,
    OSMViewer,
    ParkingMapper,
    Santa,
    RawMapEditor,
    LTN,
}

Variants

ABStreet

FifteenMin

OSMViewer

ParkingMapper

Santa

RawMapEditor

LTN

Implementations

impl Executable

pub fn replace_process<A: AppLike + 'static>(
self,
ctx: &mut EventCtx<'_>,
app: &A,
args: Vec<&str>
) -> Transition<A>

Run the given executable with some arguments. On Mac and Linux, this replaces the current process. On Windows, this launches a new child process and leaves the current alone. On web, this makes the browser go to a new page.