Module map_gui::tools::labels

Structs

DrawRoadLabels

Labels roads when unzoomed. Label size and frequency depends on the zoom level.

DrawSimpleRoadLabels

Draws labels in map-space that roughly fit on the roads. Don’t change behavior during zooming; labels are only meant to be legible when zoomed in.

Statics

SIMPLIFY_PATTERNS 🔒

Functions

cheaply_overestimate_bounds 🔒

simplify_name 🔒

simplify_patterns 🔒