Struct map_gui::tools::waypoints::InputWaypoints

pub struct InputWaypoints {
    waypoints: Vec<Waypoint>,
    snap_to_main_endpts: FindClosest<TripEndpoint>,
    snap_to_road_endpts: FindClosest<LaneID>,
    max_waypts: Option<usize>,
}

Click to add waypoints, drag them, see the list on a panel and delete them. The caller owns the Panel and the World, since there’s probably more stuff there too.

Fields
waypoints: Vec<Waypoint>snap_to_main_endpts: FindClosest<TripEndpoint>snap_to_road_endpts: FindClosest<LaneID>max_waypts: Option<usize>
Implementations

impl InputWaypoints

pub fn new( app: &dyn AppLike, snap_to_lanes_for: Vec<PathConstraints> ) -> InputWaypoints

Allows any number of waypoints

pub fn new_max_2( app: &dyn AppLike, snap_to_lanes_for: Vec<PathConstraints> ) -> Self

Only allow drawing routes with 2 waypoints. If a route is loaded with more than that, it can be modified.

pub fn overwrite(&mut self, app: &dyn AppLike, waypoints: Vec<TripEndpoint>)

The caller should call rebuild_world after this

pub fn get_panel_widget(&self, ctx: &mut EventCtx<'_>) -> Widget

pub fn get_waypoints(&self) -> Vec<TripEndpoint>

pub fn len(&self) -> usize

pub fn event( &mut self, app: &dyn AppLike, panel_outcome: Outcome, world_outcome: WorldOutcome<WaypointID> ) -> bool

If the outcome from the panel or world isn’t used by the caller, pass it along here. When this returns true, something has changed, so the caller may want to update their view of the route and call get_panel_widget and rebuild_world again.

fn snap(&self, app: &dyn AppLike, cursor: Pt2D) -> Option<TripEndpoint>

pub fn get_waypoint_color(&self, idx: usize) -> Color

pub fn rebuild_world<T: ObjectID, F: Fn(WaypointID) -> T>( &self, ctx: &mut EventCtx<'_>, world: &mut World<T>, wrap_id: F, zorder: usize )

The caller is responsible for calling initialize_hover and rebuilt_during_drag.