Crate map_gui

This crate contains common code for applications that draw and interact with a Map.

Modules

colors

A color scheme groups colors used for different map, dynamic, and UI elements in one place, to encourage deduplication. The player can also switch between different color schemes.

load

options

render

Render static map elements.

simple_app 🔒

tools

Assorted tools and UI states that’re useful for applications built to display maps.

Structs

SimpleApp

Simple app state that just renders a static map, without any dynamic agents on the map.

SimpleAppArgs

Enums

ID

Traits

AppLike

An application wishing to use the tools in this crate has to implement this on the struct that implements widgetry::SharedAppState, so that the tools here can access the map. See SimpleApp for an example implementation.