Struct map_gui::SimpleApp

pub struct SimpleApp<T> {
    pub map: Map,
    pub draw_map: DrawMap,
    pub cs: ColorScheme,
    pub opts: Options,
    pub current_selection: Option<ID>,
    pub session: T,
    pub time: Time,
}

Simple app state that just renders a static map, without any dynamic agents on the map.

Fields
map: Mapdraw_map: DrawMapcs: ColorSchemeopts: Optionscurrent_selection: Option<ID>session: T

Custom per-app state can be stored here

time: Time

If desired, this can be advanced to render traffic signals changing.

Implementations

impl<T> SimpleApp<T>

pub fn new(ctx: &mut EventCtx<'_>, args: CmdArgs, session: T) -> SimpleApp<T>

pub fn new_with_opts(
ctx: &mut EventCtx<'_>,
args: CmdArgs,
opts: Options,
session: T
) -> SimpleApp<T>

pub fn draw_unzoomed(&self, g: &mut GfxCtx<'_>)

pub fn draw_zoomed(&self, g: &mut GfxCtx<'_>, opts: DrawOptions)

pub fn recalculate_current_selection(&mut self, ctx: &EventCtx<'_>)

Assumes some defaults.

pub fn mouseover_unzoomed_roads_and_intersections(
&self,
ctx: &EventCtx<'_>
) -> Option<ID>

pub fn mouseover_unzoomed_buildings(&self, ctx: &EventCtx<'_>) -> Option<ID>

Only select buildings, and work whether zoomed in or not.

fn calculate_current_selection(
&self,
ctx: &EventCtx<'_>,
unzoomed_roads_and_intersections: bool,
unzoomed_buildings: bool
) -> Option<ID>